+++ src/entities/session/model/types.ts
@@ -4,9 +4,3 @@ export type Session = {
   accessToken: string
   userId: SessionUserId
 }
-
-// TODO: FSD: Move user to entities/user/model/types.ts
-export type User = {
-  id: Id
-  email: Email
-}
