+++ src/entities/user/api/types.ts
@@ -1,0 +1,4 @@
+export type UserDto = {
+  id: number
+  email: Email
+}
