+++ src/entities/user/lib/mapUser.ts
@@ -1,9 +1,9 @@
 import { type UserDto } from '../api/types'
-import { type User } from '../model/types'
+import { type User, type UserId } from '../model/types'
 
 export function mapUser(dto: UserDto): User {
   return {
-    id: dto.id,
+    id: dto.id as UserId,
     email: dto.email,
   }
 }
