+++ src/entities/user/model/types.ts
@@ -1,0 +1,6 @@
+export type UserId = Brand<Id, 'UserId'>
+
+export type User = {
+  id: UserId
+  email: Email
+}
