+++ src/features/authentication/logout/model/logout.ts
@@ -1,8 +1,9 @@
 import { createAsyncThunk } from '@reduxjs/toolkit'
 import { cartApi, clearCartData } from '@/entities/cart'
 import { sessionApi, clearSessionData } from '@/entities/session'
+import { userApi } from '@/entities/user/api/userApi'
 import { wishlistApi, clearWishlistData } from '@/entities/wishlist'
-import { SESSION_TAG, WISHLIST_TAG, CART_TAG } from '@/shared/api'
+import { SESSION_TAG, WISHLIST_TAG, CART_TAG, USER_TAG } from '@/shared/api'
 import { wait } from '@/shared/lib'
 
 export const logoutThunk = createAsyncThunk<void, void, { state: RootState }>(
@@ -21,6 +22,7 @@ export const logoutThunk = createAsyncThunk<void, void, { state: RootState }>(
     // dispatch(sessionApi.util.resetApiState())
 
     dispatch(sessionApi.util.invalidateTags([SESSION_TAG]))
+    dispatch(userApi.util.invalidateTags([USER_TAG]))
     dispatch(wishlistApi.util.invalidateTags([WISHLIST_TAG]))
     dispatch(cartApi.util.invalidateTags([CART_TAG]))
   }
