+++ src/shared/api/index.ts
@@ -1,4 +1,4 @@
 export { baseApi } from './baseApi'
-export { SESSION_TAG, WISHLIST_TAG, CART_TAG } from './tags'
+export { SESSION_TAG, WISHLIST_TAG, USER_TAG, CART_TAG } from './tags'
 export { invalidateAccessToken } from './invalidateTokenEvent'
 export { isFetchBaseQueryError } from './isFetchBaseQueryError'
