+++ src/shared/api/tags.ts
@@ -1,3 +1,4 @@
 export const SESSION_TAG = 'SESSION_TAG'
 export const WISHLIST_TAG = 'WISHLIST_TAG'
 export const CART_TAG = 'CART_TAG'
+export const USER_TAG = 'USER_TAG'
