+++ src/widgets/LayoutProfileCard/LayoutProfileCard.tsx
@@ -2,7 +2,8 @@ import { skipToken } from '@reduxjs/toolkit/dist/query'
 import cn from 'classnames'
 import { Link } from 'react-router-dom'
 import { selectTotalQuantity } from '@/entities/cart'
-import { selectIsAuthorized, useMeQuery } from '@/entities/session'
+import { selectIsAuthorized } from '@/entities/session'
+import { useMeQuery } from '@/entities/user'
 import { selectProductIdsInWishlist } from '@/entities/wishlist'
 import { LogoutButton } from '@/features/authentication/logout'
 import { useFeatureSlicedDebug } from '@/shared/lib'
